delivery service completed

+++ Frontend/AccoMart/src/app/pages/delivery-services/delivery-services.component.ts
@@ -5,18 +5,37 @@ import { deliveryServices } from '../../services/delivery.service';
 import { DeliveryService } from '../../interfaces/deliveryService';
 import { CommonModule } from '@angular/common';
 import { createDeliveryService } from '../../interfaces/createDeliveryService';
+import { FormsModule } from '@angular/forms'; // Import FormsModule
 
 @Component({
   selector: 'app-delivery-services',
   standalone: true,
-  imports: [SidebarComponent,NavbarComponent,CommonModule],
+  imports: [SidebarComponent,NavbarComponent,CommonModule,FormsModule],
   templateUrl: './delivery-services.component.html',
   styleUrl: './delivery-services.component.css'
 })
 export class DeliveryServicesComponent {
+
   constructor(private deliveryService: deliveryServices) {}
   deliveryServicesList: DeliveryService[] = [];
+  openAddServicePopup:boolean = false;
+  openEditServicePopup:boolean = false;
+  editServiceId:number;
+  
+  serviceToAdd: createDeliveryService = {
+    imageUrl: '',
+    serviceName: '',
+    price: 0,
+    deliveryDays: 0
+  };
 
+  serviceToEdit: createDeliveryService = {
+    imageUrl: '',
+    serviceName: '',
+    price: 0,
+    deliveryDays: 0
+  };
+  
   ngOnInit(): void {
     this.fetchDeliveryServices();
   }
@@ -26,23 +45,27 @@ export class DeliveryServicesComponent {
     }).catch(error => {
       console.error('Error fetching delivery services:', error);
     });
+
+    
   }
 
   deleteDeliveryService(id: number) {
     
-    this.deliveryService.deleteDeliveryService(id).then(() => {
-
-      // After successful deletion, fetch the updated list of delivery services
-      this.fetchDeliveryServices();
-    }).catch(error => {
-      console.error('Error deleting delivery service:', error);
-    });
+    if(confirm("Are you sure you want to delete this Service?"))
+      {
+        this.deliveryService.deleteDeliveryService(id).then(() => {
+          this.fetchDeliveryServices();
+        }).catch(error => {
+          console.error('Error deleting delivery service:', error);
+        });
+      }
   }
 
 
-  editDeliveryService(deliveryService: createDeliveryService,id:number) {
-    this.deliveryService.editDeliveryService(deliveryService,id).then(() => {
-      console.log("Delivery Service succesfully updated")
+  editDeliveryService(deliveryService: createDeliveryService) {
+    this.deliveryService.editDeliveryService(deliveryService,this.editServiceId).then(() => {
+      console.log("Delivery Service succesfully updated");
+      this.openAddServicePopup= false;
       this.fetchDeliveryServices();
     }).catch(error => {
       console.error('Error editing delivery service:', error);
@@ -52,9 +75,37 @@ export class DeliveryServicesComponent {
   createDeliveryService(newDeliveryService: createDeliveryService) {
     this.deliveryService.addDeliveryService(newDeliveryService).then(() => {
       // After successful creation, fetch the updated list of delivery services
+      this.openAddServicePopup =false;
       this.fetchDeliveryServices();
     }).catch(error => {
       console.error('Error creating delivery service:', error);
     });
   }
+
+  openAddPopup()
+  {
+    this.openAddServicePopup = true;
+  }
+
+  openEditPopup(Id:number,service:createDeliveryService)
+  {
+    this.editServiceId = Id;
+    this.openAddServicePopup = true;
+    this.serviceToEdit = {...service};
+
+  }
+
+  
+  closeAddPopup()
+  {
+    this.openAddServicePopup = false;
+  }
+
+  closeEditPopup()
+  {
+    this.openAddServicePopup = false;
+  }
+
 }
+
+
